fix dynamic server usage
